LUCENE-10155: Refactor TestMultiMMap into a BaseChunkedDirectoryTestCase
BaseChunkedDirectoryTestCase is an extension of BaseDirectoryTestCase
where the concrete test class instantiates with a specified chunk size.
It then tries to test boundary conditions around all the chunking.
